+++ src/func/java/org/smoothbuild/builtin/java/javac/PackagedJavaFileObjects.java
@@ -5,7 +5,7 @@ import static org.smoothbuild.builtin.java.util.JavaNaming.isClassFilePredicate;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.smoothbuild.builtin.java.Unjarer;
+import org.smoothbuild.builtin.compress.UnzipFunction;
 import org.smoothbuild.lang.message.ErrorMessage;
 import org.smoothbuild.lang.plugin.Container;
 import org.smoothbuild.lang.value.Array;
@@ -15,11 +15,9 @@ import org.smoothbuild.lang.value.SFile;
 public class PackagedJavaFileObjects {
   public static Iterable<InputClassFile> classesFromJars(Container container,
       Iterable<Blob> libraryJars) {
-    Unjarer unjarer = new Unjarer(container);
     Set<InputClassFile> result = new HashSet<>();
-
     for (Blob jarBlob : libraryJars) {
-      Array<SFile> files = unjarer.unjar(jarBlob, isClassFilePredicate());
+      Array<SFile> files = UnzipFunction.unzip(container, jarBlob, isClassFilePredicate());
       for (SFile classFile : files) {
         InputClassFile inputClassFile = new InputClassFile(classFile);
         if (result.contains(inputClassFile)) {
@@ -30,7 +28,6 @@ public class PackagedJavaFileObjects {
         }
       }
     }
-
     return result;
   }
 }